refactor: 미사용 코드 정리
- 빈 파일 및 스텁 파일 정리
- deprecated 네비게이션 함수 제거 및 urlUtils로 이동
